[#177145461] use new CardLogoPreview component to favor composition
